+++ packages/ssml-editor-react/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @ssml-builder/ssml-editor-react
+# @ssml-builder-js/ssml-editor-react
 
 ## 1.0.0
 
@@ -9,4 +9,4 @@
 ### Patch Changes
 
 - Updated dependencies [6c78e70]
-  - @ssml-builder/ssml-core@1.0.0
+  - @ssml-builder-js/ssml-core@1.0.0